# terrarium-core/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Экстенсивная (excessive) документация API системы управления микроклиматом террариума на базе Raspberry Pi 5.",
        "title": "API Платформы Климат-Контроля Террариума (Terrarium Climate)",
        "contact": {},
        "version": "1.0.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/api/v1/config": {
            "get": {
                "description": "Возвращает текущие настройки террариума: полярные целевые значения температуры, влажности, гистерезиса и пороги аварийных отключений. Настройки подтягиваются из Postgres.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "System",
                    "Configuration"
                ],
                "summary": "Получить текущую конфигурацию (Настройки Автоматизации)",
                "responses": {
                    "200": {
                        "description": "Успешное получение настроек",
                        "schema": {
                            "$ref": "#/definitions/models.ConfigPayload"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка (например, сбой соединения с БД)",
                        "schema": {
                            "$ref": "#/definitions/models.HTTPError"
                        }
                    }
                }
            },
            "put": {
                "description": "Принимает новые пороговые значения (Payload) и валидирует их. В случае успеха, новые пороги сохраняются в БД.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "System",
                    "Configuration"
                ],
                "summary": "Обновить границы климатического контроля",
                "parameters": [
                    {
                        "description": "Объект новых настроек климата",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.ConfigPayload"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Конфигурация успешно обновлена",
                        "schema": {
                            "$ref": "#/definitions/models.ConfigPayload"
                        }
                    },
                    "400": {
                        "description": "Невалидный Payload",
                        "schema": {
                            "$ref": "#/definitions/models.HTTPError"
                        }
                    },
                    "500": {
                        "description": "Ошибка записи в базу данных postgres",
                        "schema": {
                            "$ref": "#/definitions/models.HTTPError"
                        }
                    }
                }
            }
        },
...
        "/api/v1/relay-logs": {
            "get": {
                "description": "Возвращает аудит-лог всех событий включения/выключения реле с причиной и временной меткой. Поддерживает пагинацию.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Logs"
                ],
                "summary": "Получить журнал переключений реле",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Количество записей (по умолчанию 50, макс 500)",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Смещение для пагинации (по умолчанию 0)",
                        "name": "offset",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Журнал переключений",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.RelayLogEntry"
                            }
                        }
                    },
                    "500": {
                        "description": "Ошибка чтения из БД",
                        "schema": {
                            "$ref": "#/definitions/models.HTTPError"
                        }
                    }
                }
            }
        },
...
        "/api/v1/relays/{id}/toggle": {
            "post": {
                "description": "Сигнализирует Raspberry Pi переключить уровень GPIO на конкретном пине. Работает только в MANUAL.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Hardware Control (Manual Mode)"
                ],
                "summary": "Переключить конкретное реле [Требует MANUAL режим]",
                "parameters": [
                    {
                        "enum": [
                            "heat_mat",
                            "fogger",
                            "light",
                            "spare"
                        ],
                        "type": "string",
                        "description": "ID реле для переключения",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Телеметрия с запросом на активацию/деактивацию",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.RelayToggleRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Реле успешно переключено",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Неизвестный ID реле",
                        "schema": {
                            "$ref": "#/definitions/models.HTTPError"
                        }
                    },
                    "403": {
                        "description": "Система находится в режиме AUTO (ручное управление запрещено)",
                        "schema": {
                            "$ref": "#/definitions/models.HTTPError"
                        }
                    }
                }
            }
        },
        "/api/v1/schedules": {
            "get": {
                "description": "Возвращает список всех расписаний автоматического включения/выключения реле по времени.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Schedules"
                ],
                "summary": "Получить все расписания реле",
                "responses": {
                    "200": {
                        "description": "Список расписаний",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Schedule"
                            }
                        }
                    },
                    "500": {
                        "description": "Ошибка чтения из БД",
                        "schema": {
                            "$ref": "#/definitions/models.HTTPError"
                        }
                    }
                }
            },
            "post": {
                "description": "Создаёт запись расписания для автоматического включения/выключения реле по времени суток.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Schedules"
                ],
                "summary": "Создать новое расписание реле",
                "parameters": [
                    {
                        "description": "Данные расписания",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.ScheduleRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Расписание успешно создано",
                        "schema": {
                            "$ref": "#/definitions/models.Schedule"
                        }
                    },
                    "400": {
                        "description": "Невалидный Payload",
                        "schema": {
                            "$ref": "#/definitions/models.HTTPError"
                        }
                    },
                    "500": {
                        "description": "Ошибка записи в БД",
                        "schema": {
                            "$ref": "#/definitions/models.HTTPError"
                        }
                    }
                }
            }
        },
        "/api/v1/schedules/{id}": {
            "put": {
                "description": "Изменяет параметры существующего расписания реле по его UUID.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Schedules"
                ],
                "summary": "Обновить существующее расписание",
                "parameters": [
                    {
                        "type": "string",
                        "description": "UUID расписания",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Обновлённые данные",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.ScheduleRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Расписание обновлено",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Невалидный Payload",
                        "schema": {
                            "$ref": "#/definitions/models.HTTPError"
                        }
                    },
                    "404": {
                        "description": "Расписание не найдено",
                        "schema": {
                            "$ref": "#/definitions/models.HTTPError"
                        }
                    },
                    "500": {
                        "description": "Ошибка обновления",
                        "schema": {
                            "$ref": "#/definitions/models.HTTPError"
                        }
                    }
                }
            },
            "delete": {
                "description": "Удаляет запись расписания по UUID.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Schedules"
                ],
                "summary": "Удалить расписание реле",
                "parameters": [
                    {
                        "type": "string",
                        "description": "UUID расписания",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Расписание удалено",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Расписание не найдено",
                        "schema": {
                            "$ref": "#/definitions/models.HTTPError"
                        }
                    }
                }
            }
        },
...
        "/api/v1/system/mode": {
            "post": {
                "description": "Позволяет пользователю полностью перехватить контроль над реле.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "System"
                ],
                "summary": "Изменить глобальный режим системы (AUTO или MANUAL)",
                "parameters": [
                    {
                        "description": "Целевой режим: 'AUTO' или 'MANUAL'",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.ModeRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Режим успешно изменен",
                        "schema": {
                            "$ref": "#/definitions/models.ModeRequest"
                        }
                    },
                    "400": {
                        "description": "Неверно заданный режим",
                        "schema": {
                            "$ref": "#/definitions/models.HTTPError"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "models.ConfigPayload": {
            "description": "Payload конфигурации для управления поведением механизма климат-контроля",
            "type": "object",
            "required": [
                "cold_max_threshold",
                "emergency_max_threshold",
                "humidity_max",
                "humidity_min",
                "hysteresis_hum",
                "hysteresis_temp",
                "warm_target_max",
                "warm_target_min"
            ],
            "properties": {
                "cold_max_threshold": {
                    "description": "Максимально допустимая температура в холодной зоне (°C). Если превышена, обогрев принудительно отключается.\nExample: 26.5",
                    "type": "number",
                    "maximum": 35,
                    "minimum": 20,
                    "example": 26.5
                },
                "emergency_max_threshold": {
                    "description": "Температурный порог теплой зоны (°C), при котором система аварийно отключает всё и шлёт Alert в Telegram.\nExample: 35.0",
                    "type": "number",
                    "maximum": 45,
                    "minimum": 30,
                    "example": 35
                },
                "humidity_max": {
                    "description": "Максимальная влажность (%), при достижении которой фоггер отключается.\nExample: 65.0",
                    "type": "number",
                    "maximum": 100,
                    "minimum": 0,
                    "example": 65
                },
                "humidity_min": {
                    "description": "Минимальная влажность (%), при которой включается фоггер (генератор тумана).\nExample: 50.0",
                    "type": "number",
                    "maximum": 100,
                    "minimum": 0,
                    "example": 50
                },
                "hysteresis_hum": {
                    "description": "Гистерезис влажности (%), для предотвращения частого срабатывания фоггера.\nExample: 2.0",
                    "type": "number",
                    "maximum": 10,
                    "minimum": 0.5,
                    "example": 2
                },
                "hysteresis_temp": {
                    "description": "Температурный гистерезис (°C), чтобы избежать \"дребезга\" реле около целевого значения.\nExample: 0.5",
                    "type": "number",
                    "maximum": 5,
                    "minimum": 0.1,
                    "example": 0.5
                },
                "warm_target_max": {
                    "description": "Максимальная целевая температура в теплой зоне (°C), при достижении которой обогрев отключается.\nОграничения: от WarmTargetMin до 40.0.\nExample: 33.0",
                    "type": "number",
                    "maximum": 40,
                    "minimum": 20,
                    "example": 33
                },
                "warm_target_min": {
                    "description": "Минимальная целевая температура в теплой зоне (°C), при которой включается обогрев.\nОграничения: от 20.0 до 40.0.\nExample: 31.5",
                    "type": "number",
                    "maximum": 40,
                    "minimum": 20,
                    "example": 31.5
                }
            }
        },
...
        "models.HTTPError": {
            "description": "Стандартный формат возвращаемой ошибки при нештатных или невалидных запросах",
            "type": "object",
            "properties": {
                "code": {
                    "description": "HTTP код статуса ошибки\nExample: 400",
                    "type": "integer",
                    "example": 400
                },
                "message": {
                    "description": "Текстовое описание ошибки или причина отказа\nExample: \"Invalid request payload\"",
                    "type": "string",
                    "example": "Параметры выходят за допустимые пределы"
                }
            }
        },
        "models.ModeRequest": {
            "description": "Запрос для переключения между АВТОМАТИЧЕСКОЙ и РУЧНОЙ работой механизмов.",
            "type": "object",
            "required": [
                "mode"
            ],
            "properties": {
                "mode": {
                    "description": "Целевой режим работы системы. Допускается 'AUTO' (автоматика) или 'MANUAL' (ручное управление).\nExample: MANUAL",
                    "type": "string",
                    "enum": [
                        "AUTO",
                        "MANUAL"
                    ],
                    "example": "MANUAL"
                }
            }
        },
...
        "models.RelayToggleRequest": {
            "description": "Запрос на принудительное переключение состояния реле",
            "type": "object",
            "properties": {
                "state": {
                    "description": "Желаемое состояние реле (true для включения, false для выключения)\nExample: true",
                    "type": "boolean",
                    "example": true
                }
            }
        },
...
    }
}